requirements.txt
albumentations==1.0.1
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
imutils==0.5.4
matplotlib==3.4.2
numpy==1.21.0
opencv-python-headless==4.5.1.48
pandas==1.3.0
Pillow==8.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2021.7.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0